Modify type

+++ matcron/lib/app/features/type/domain/entities/mattress_type.dart
@@ -9,7 +9,7 @@ class MattressTypeEntity {
   String? recyclingDetails;
   double? expectedLifespan;
   double? warrantyPeriod;
-  double? stock;
+  int? stock;
 
   MattressTypeEntity({
     this.id,
